Merge staging to development

+++ packages/contentstack-audit/src/modules/workflows.ts
@@ -57,6 +57,7 @@ export default class Workflows {
       return {};
     }
 
+
     this.workflowPath = join(this.folderPath, this.fileName);
     this.workflowSchema = existsSync(this.workflowPath)
       ? values(JSON.parse(readFileSync(this.workflowPath, 'utf8')) as Workflow[])
@@ -87,7 +88,7 @@ export default class Workflows {
       this.log(
         $t(auditMsg.SCAN_WF_SUCCESS_MSG, {
           name: workflow.name,
-          module: this.config.moduleConfig[this.moduleName].name,
+          uid: workflow.uid,
         }),
         'info',
       );
@@ -144,9 +145,9 @@ export default class Workflows {
 
   async writeFixContent(newWorkflowSchema: Record<string, Workflow>) {
     if (
-      this.fix ||
-      (!(this.config.flags['copy-dir'] || this.config.flags['external-config']?.skipConfirm) &&
-        (this.config.flags.yes || (await ux.confirm(commonMsg.FIX_CONFIRMATION))))
+      this.fix &&
+      !(this.config.flags['copy-dir'] || this.config.flags['external-config']?.skipConfirm) &&
+      (this.config.flags.yes || (await ux.confirm(commonMsg.FIX_CONFIRMATION)))
     ) {
       writeFileSync(
         join(this.folderPath, this.config.moduleConfig[this.moduleName].fileName),
